Added additional assets, updated the tilesheet. Added a new component CDestroyable which can be added to any tile which we want the player to be able to destroy. This can be used to reveal hidden parts of the map for example, allowing for exploration.
Updated the LevelEditor and the PlayScene. PlayScene now checks to see if a tile is destroyable or not...if not, only the bullet is destroyed upon collision. Level Editor now has a green outline displayed around any tile which has the CDestroyable component. I'm realizing that in the future the game would need a massive overhaul of the LevelEditing system (also, why am I using .txt and not .json???), but for now it does what I need it to do.